Rename `loop.rb` to `interval_loop.rb` to match module name
