+++ lib/ddtrace/workers/interval_loop.rb
